input_requirements.txt
flask
uuid
python-dotenv
pyaudio
wave
pydub
googletrans
langdetect
google-cloud-translate
google-cloud-speech
flask_session